requirements.txt
httpx==0.24.0
Pillow==9.5.0
playwright~=1.43.0
tenacity==8.2.2
tornado~=6.4
PyExecJS==1.5.1
opencv-python==4.7.0.72
aiomysql==0.2.0
redis~=4.6.0
pydantic==2.5.2
aiofiles~=23.2.1

numpy~=1.24.4
torch~=2.2.1+cu118
torchvision~=0.17.1+cu118